Add custom generator support to open api spec
