+++ tests/Unit/OpenAPISpecWriterTest.php
@@ -7,6 +7,7 @@ use Illuminate\Support\Arr;
 use Knuckles\Camel\Camel;
 use Knuckles\Camel\Output\OutputEndpointData;
 use Knuckles\Scribe\Tests\BaseUnitTest;
+use Knuckles\Scribe\Tests\Fixtures\TestOpenApiGenerator;
 use Knuckles\Scribe\Tools\DocumentationConfig;
 use Knuckles\Scribe\Writing\OpenAPISpecWriter;
 
@@ -808,6 +809,30 @@ class OpenAPISpecWriterTest extends BaseUnitTest
                 ],
             ],
         ], $results['paths']['/path']['post']['responses']);
+    }
+
+    /** @test */
+    public function can_extend_openapi_generator()
+    {
+        $endpointData1 = $this->createMockEndpointData([
+            'uri' => '/path',
+            'httpMethods' => ['POST'],
+            'custom' => ['permissions' => ['post:view']]
+        ]);
+        $groups = [$this->createGroup([$endpointData1])];
+        $extraGenerator = TestOpenApiGenerator::class;
+        $config = array_merge($this->config, [
+            'openapi' => [
+                'generators' => [
+                    $extraGenerator,
+                ],
+            ],
+        ]);
+        $writer = new OpenAPISpecWriter(new DocumentationConfig($config));
+
+        $results = $writer->generateSpecContent($groups);
+
+        $this->assertEquals([['default' => ['post:view']]], $results['paths']['/path']['post']['security']);
     }
 
     protected function createMockEndpointData(array $custom = []): OutputEndpointData
